[CI] Split test_pp_single_node into base and extra suites
